Move find_start_end to tristan.data
