+++ src/tristan/data.py
@@ -199,6 +199,29 @@ def cue_times(
     return da.unique(data[cue_time_key][index].values)
 
 
+def find_start_end(data: dd.DataFrame) -> tuple[int, int]:
+    """
+    Find the shutter open and shutter close timestamps.
+
+    Args:
+        data:  LATRD data.  Must contain one 'cue_id' entry and one
+               'cue_timestamp_zero' entry.  The two arrays are assumed to have the
+               same length.
+
+    Returns:
+        The shutter open and shutter close timestamps, in clock cycles.
+    """
+    selection = (data[cue_id_key] == shutter_open) | (data[cue_id_key] == shutter_close)
+    shutter_times = data[selection]
+    print("Finding detector shutter open and close times.")
+    (shutter_times,) = compute_with_progress(shutter_times, gather=True)
+    shutter_times.drop_duplicates(inplace=True)
+    shutter_times.set_index(cue_id_key, inplace=True)
+    start, end = shutter_times.loc[[shutter_open, shutter_close]].values
+
+    return start, end
+
+
 def seconds(timestamp: int, reference: int = 0) -> Quantity:
     """
     Convert a Tristan timestamp to seconds, measured from a given reference timestamp.
